Finalizing index.html and style.css
SUMMARY: Perfectioning the stye with minor changes to index.html, to approach the look of the mockups.
MEMO: Provide a class to the divs containing the text withing the "New Members" and "Recent Activity" sections to center the text through the display flex property.
NEXT: Finalize repsonsiveness, and Work on Chart.js to add the charts.